requirements.txt
amqp==2.5.2
bcrypt==3.1.7
billiard==3.6.1.0
boto3==1.10.28
botocore==1.13.28
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.7
django-celery-beat==1.5.0
django-filter==2.2.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.4
django-timezone-field==3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
fabric2==2.5.0
idna==2.8
importlib-metadata==0.23
invoke==1.3.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==7.2.0
openapi-codec==1.3.2
paramiko==2.6.0
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
python-crontab==2.4.0
python-dateutil==2.8.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
s3transfer==0.2.1
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
